refactor: refatora projeto
