+++ .env.example
@@ -1,12 +1,12 @@
-DOMAIN_URL
-DATABASE_URL
-DB_USER
-DB_PASSWORD
-CACHE_URL
-CACHE_PASSWORD
-PORT
-NODE_ENV
-RATE_LIMIT_FIXED_WINDOW_IN_MINUTES
-RATE_LIMIT_TOKEN
+DOMAIN_URL=http://localhost:3001/v1
+DATABASE_URL=postgres://postgres:postgres@localhost:5432/ShortenedUrls?sslmode=disable
+DB_USER=postgres
+DB_PASSWORD=postgres
+CACHE_URL=
+CACHE_PASSWORD=
+PORT=3001
+NODE_ENV=development
+RATE_LIMIT_FIXED_WINDOW_IN_MINUTES=90
+RATE_LIMIT_TOKEN=1000
 DOCKERHUB_USERNAME
 DOCKERHUB_IMAGE
